app/pip-requirements.txt
# Requirements
django
channel
channels_redis
pytest
django-registration
psycopg2
websocket_client
